Minor fixes to the build infrastructure.
Build packet failed because SRCROOT was not exported.

SSHD seemed to fail to start up because the passwd file got overwritten
during the add_files step, and opkg didn't rerun installation for openssh
because it thought it was already done. The fix was to not recopy skel to
apps if it has already happened.

I also wanted to start saving the logs of the autoclient.

+++ tasks/build/add_files.sh
@@ -75,10 +75,12 @@ fi
 
 ##
 ## Copy the skeleton over so the proper environment is set up for postinst
-## scripts.
+## scripts. Don't do this if it has already been done.
 ##
 
-cp -Rpf $BINROOT/skel/* $BINROOT/apps/ || true
+if ! [ -f $BINROOT/apps/etc/inittab ]; then
+    cp -Rpf $BINROOT/skel/* $BINROOT/apps/ || true
+fi
 
 ##
 ## Copy the script that automatically loads the Python build client.
